requirements.txt
--trusted-host git.fenrys.io
-i http://git.fenrys.io:3141/root/pypi/+simple
abstract-queue==1.0
aniso8601==8.0.0
appdirs==1.4.3
apscheduler==3.6.3
attrs==19.3.0
bandit==1.6.2
black==18.9b0
certifi==2019.11.28
click==7.0
coverage==5.0.3
ddd-es==1.3.3
dependency-injector==3.15.6
entrypoints==0.3
faker==4.0.1
fakeredis==1.2.1
flake8-docstrings==1.5.0
flake8==3.7.9
flask-httpauth==3.3.0
flask-restful==0.3.8
flask==1.1.1
gitdb==4.0.2
gitpython==3.1.0
gunicorn==20.0.4
importlib-metadata==1.5.0 ; python_version < '3.8'
isort==4.3.21
itsdangerous==1.1.0
jinja2==2.11.1
loguru==0.4.1
markupsafe==1.1.1
mccabe==0.6.1
mongomock==3.19.0
more-itertools==8.2.0
morfessor==2.0.6
packaging==20.1
pathlib2==2.3.5
pbr==5.4.4
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pydocstyle==5.0.2
pyflakes==2.1.1
pygrate==0.0.3
pymongo==3.10.1
pymysql==0.9.3
pyparsing==2.4.6
pytest-cov==2.8.1
pytest-flask==1.0.0
pytest-mock==2.0.0
pytest==5.3.5
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
pyyaml==5.3
redis==3.4.1
rope==0.16.0
sentinels==1.0.0
sentry-sdk==0.6.9
six==1.14.0
smmap==3.0.1
snowballstemmer==2.0.0
sortedcontainers==2.1.0
stevedore==1.32.0
text-unidecode==1.3
toml==0.10.0
tzlocal==2.0.0
unidecode==1.1.1
urllib3==1.25.8
vulture==1.3
wcwidth==0.1.8
werkzeug==1.0.0
zipp==3.1.0
